Extract accounting cash-flow helpers into a partial
Increment 2e of the endpoint-monolith decomposition. Moves the accounting
cash-flow helper cluster out of WorkstationEndpoints.cs into a new sibling partial
WorkstationEndpoints.AccountingCashFlow.cs.

Moved (4 whole private static methods, closed cluster): BuildAccountingWorkspaceCashFlowSummary,
BuildAccountingRunCashFlowSummary, GetLedgerCashBalance, GetCashVariance. The two public
builders are called from accounting endpoints that remain in core (cross-partial calls);
the two ledger-cash helpers are cluster-internal. Behavior-preserving relocation within the
single partial class. Core drops 6865 -> 6798 lines; ratchet baseline retightened.
